// frontend/src/api/models/ItemSummary.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ItemSummary = {
    /**
     * Unique identifier for the item
     */
    itemId: string;
    /**
     * Name of the item
     */
    name: string;
    /**
     * Price of the item in the default currency
     */
    price: number;
};