// frontend/src/api/models/ItemDetail.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ItemDetail = {
    /**
     * Unique identifier for the item
     */
    itemId: string;
    /**
     * Name of the item
     */
    name: string;
    /**
     * Price of the item in the default currency
     */
    price: number;
    /**
     * Full description of the item
     */
    description: string;
    /**
     * Category of the item
     */
    category?: string;
    /**
     * Brand name
     */
    brand?: string;
    /**
     * Stock Keeping Unit identifier
     */
    sku?: string;
};